Move bevy shim to own crate
