--- matchbox_socket/src/integrations/mod.rs
@@ -1,10 +1,0 @@
-use cfg_if::cfg_if;
-
-cfg_if! {
-    if #[cfg(feature = "bevy-plugin")] {
-        mod bevy_plugin;
-        pub use bevy_plugin::*;
-    }
-}
-#[cfg(feature = "ggrs-socket")]
-mod ggrs_socket;
